🔍 Refactor `/request` dto
